+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 2.29.0
+### Breaking
+* remove undocumented `<large>`/`<big>`/`<wide>` and add `<font>` tag to XML formatter
+### Feature
+* add animation to `exec` and `execAnimation` option [#679](https://github.com/jcubic/jquery.terminal/issues/679)
+* add warning when calling invoke_key on disabled terminal [#678](https://github.com/jcubic/jquery.terminal/issues/678)
+* if interpreter returns a promise and it's animating the terminal will not pause (for exec array)
+* allow to use brackets inside extended commands (e.g. JavaScript code to hide commands)
+* expose `$.terminal.xml_formatter.tags`
+* add `--glow` [conditional hack](https://www.kizu.ru/conditions-for-css-variables/)
+### Bugfix
+* tweak types for cmd function prompt
+* fix broken order of execution in `exec` when using pause/resume
+* fix mixing extended commands and terminal formatting
+* fix invoking `get_command()` in interpreter (clear command after user action)
+* fix parse_command when command have leading spaces [#681](https://github.com/jcubic/jquery.terminal/issues/681)
+
 ## 2.28.1
 ### Bugfix
 * fix applying style of text-decoration (e.g. underline)